Apply Google Page Speed recommendations
