+++ components/Home/index.js
@@ -1,3 +1,3 @@
-export { default as Gif } from './Gif'
+export { default as Video } from './Video'
 export { default as Description } from './Description'
 export { default as Links } from './Links/Links'
